+++ core/web/app/pages/_app.js
@@ -1,0 +1,6 @@
+// core/web/app/pages/_app.js
+import '../styles/globals.css';
+
+export default function App({ Component, pageProps }) {
+  return <Component {...pageProps} />;
+}